requirements.txt
numpy==1.14.3
Keras==1.2.2
matplotlib==2.0.0
tensorflow_gpu==1.4.0
Theano==1.0.0
six==1.11.0
Pillow==5.2.0
scipy==1.1.0
scikit_learn==0.19.2
tensorflow==1.10.1
